Bug fixes. Validation. Feature added.
Moved from the $.change() function to an onClose even on the
datetimepicker widget to avoid a bug where the $.change() event was
called multiple times.

Users can no longer just click the play button after the page has
loaded. In the future I may implement this being able to start playback
from the beginning on whatever is on the timeine. For now this is just
to stop a bug occuring.

Stopped users from being able to select a Range Start after the Range
End and the Range End before the Range Start. When the user tries to do
this now they get an error message and a non-offending date is
automatically selected +/- 1 day.

On the "Choose Cameras" dialog if the user clicks a camera which has
already been added it is removed from the page.