Added features. Usability tweaks.
Global AJAX error handler added.
Notifications now displayed when the timeline loads new events
successfully or upon error.
When the date range is changed when no cameras are selected the timeline
is wiped and the user is told as much.
Show All and Hide All links added to the Add Cameras dialog.
Clicking the Show Cameras button will toggle the dialog.