Features added. Bug fix.
Each camera now has a coloured box in the bar above the video next to
the close icon. This can be seen as a key of sorts for the timeline
events which are coloured in the same style.

Bugs fixed which arose from adding the box and changing some elements
from <a> tags to <button> tags.

Fixed a bug that stopped the Show All cameras button from working
because it was adding the item to the chosencameras array then passing
to the addMonitor function which performed the same check.