Bug fixes. Removed old code that's no longer needed.
* ajaxRequests array added to store all currently active ajax request
instances.
* haschosencameras boolean variable added to detect if the user has
picked a preset or added any cameras. Used to stop the initial timeline
(which has the last 24 hours of data on all cameras on) being requeried.
* Ajax error notification removed when jqXHR.status code 0. This is now
silently logged to the console as the error can be thrown by
interrupting an ajax request.
* Fixed a bug whereby everytime an event started playing, a div would be
injected. Fixed by removing parseInt(monitorId)
* Fixed a bug which meant multiple camera instances could be added by
spamming the preset links.