Bugfixes. Features added.
* stopLiveStreams() function added for debug purposes

* Bootstrap grid classes no longer appended when a new camera is added
as flexbox is now used instead \0/

* rel="tooltip" removed in favour of data-rel="tooltip" to pass HTML5
validation checks

* Switching between Live View and Playback mode now disabled some UI
controls to make it clearer to the user that they can't be used.
Functions still check the mode though to ensure nothing goes wrong.

* Bug fix: Current date / time reset when entering live view.

* Feature added: Scale increase/decrease and reset buttons added to
control the size of the cameras to ensure that the cameras fit in the
pane. (Should be handled via Flexbox but it's bugged at the moment)

* Added col-container div wrapper to camera instances therefore I've
updated some of the function calls to use
parent().parent().parent().parent() instead of
parent().parent().parent()

* Feature added: Export button on the Playback (frontend) is now
enabled! Spawns a colorbox window for the Events backend with relevant
data already loaded.