v0.7.75 - More notifications added. $.deferred used.
* [0.7.75] - Notifications added to let the user know when presets / all
cameras / individual cameras have been loaded. The notifications are
wrapped in $.deferred functions $.when and $.then to ensure they are
shown at the right time to the user. Login screen is no longer dark.
Notifications also added when changing the camera sizes with the buttons
at the top of the screen. This was needed as sometimes  when resetting
the size it was hard to tell if any change had actually been done.