v0.7.77 - Bug fix. Fullscreen toggle button added.
* [0.7.76] - Bug fix: When resizing the browser window whilst viewing a
camera fullscreen the modal window wouldn't resize. This has now been
resolved and the expected behaviour occurs.
* [0.7.77] - Minor feature: Fullscreen toggle button added in the bottom
right of the playback view. If the browser supports the HTML5 fullscreen
API it can be used to hide the browser chrome. If not a message is
displayed telling the user to use F11 instead.