v0.7.82 - Bug fixes. Minor feature release. Code cleanup.
* [0.7.80] - Minor feature relase: "Load Events" button added for use
when changing the playback range. This change serves two purposes - 1)
clearer UX. When shifting between weeks it was particularly fiddly to
change the start then the end range to the time period you were looking
for. 2) It halves the number of requests to the server when changing
ranges.
* [0.7.81] - Code cleanup: Removed old code on date pickers that was
commented out until new code was tested. "Done" button on date pickers
now reads "Hide".
* [0.7.82] - Bug fix: Changing playback speed whilst playing an event
was broken. This was down to clearPlayback() wiping the current event
variable. This has been solved by storing the "currentevent" in a local
variable for use when restarting the current event at the desired
playback speed. Timeline height has been bumped up and borders have been
tweaked for asthetics.