v0.7.84 - Hot fix for Firefox. Bug fix. Styling changes.
* [0.7.83] - Hot fix for Firefox - Live view streams now actually
playback properly. This was caused by setting backface-visbility in the
CSS which affected the rendering of live streams. This is now toggled
between live view and playback mode via playback.js.
* [0.7.84] - Bug fix: Fixed a bug where some cameras wouldn't appear on
the page if they were setup to only be available for live view (e.g.
they don't store events). This has now been fixed and the event count
from the top of each monitor has been removed as it was only ever a
static value on page load and did not change with the timeline.