v0.7.96 - Major bug fixes
[0.7.93] - Bug fix: User list was showing N/A for user presets due to an
error when checking an array. This has now been resolved. If a user has
set a preset it will show the preset name in the table, otherwise "N/A"
will be shown.
[0.7.94] - Bug fix: The login page would intermittently require you to
login twice. This was due to a bug with the code that checked for an
incorrect password. This has been resolved but IS INSECURE for the time
being - this will be addressed in a future release. Feature: When your
session times out and you were on a specific page, after re-loggining in
you would be dumped back to playback. The system now will redirect you
back to the page you were on before you timed out.
[0.7.95] - Bug fix: The camera status icons shown in the "Cameras"
dropdown menu were incorrect. They would show an online status when
therew as actually a problem with the source. This has been resolved and
new font awesome icons are in use.
[0.7.96] - Bug fix: When deleting a preset any users who had that preset
assigned to them would be left in that state causing problems in
playback mode. Now when a preset is deleted any users which had that
preset assigned to them are updated to show "All Cameras" as their
default preset. Minor tweak: The admin information table is now striped.